Add files via upload
Extract caridac time series signal from the data. show the plot
Extract Respiratory time series data. Show the plot.
Filter the signals and find the different bands corresponding to the mentioned tasks.
Extimate the power spectrum of the and lable the compoenets corresponding to activitiies.
Detend the signal
Extract P300 or ERP from the signals based on the type of task.